moving components and all helpers from Dashboard to Map
